# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Modelo que contiene la información del tiempo extra en un partido",
        "contact": {}
    },
    "paths": {
        "/api/matches": {
            "get": {
                "description": "Retorna una lista con todos los partidos registrados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "Obtener todos los partidos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Match"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Crea un nuevo registro de partido con los datos básicos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "Crear un nuevo partido",
                "parameters": [
                    {
                        "description": "Datos del partido",
                        "name": "match",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Match"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Match"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/matches/{id}": {
            "get": {
                "description": "Retorna los datos de un partido específico",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "Obtener partido por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del partido",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Match"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Modifica los datos de un partido existente por ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "Actualizar partido",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del partido",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos actualizados",
                        "name": "match",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.Match"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Match"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Elimina un partido de la base de datos por ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "Eliminar partido",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del partido",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sin contenido",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/matches/{id}/extratime": {
            "patch": {
                "description": "Establece el valor de tiempo extra en un partido específico",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matches"
                ],
                "summary": "Establecer tiempo extra",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del partido",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tiempo extra en formato MM:SS",
                        "name": "extra_time",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ExtraTimePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "main.ExtraTimePayload": {
            "description": "Modelo que contiene la información del tiempo extra en un partido",
            "type": "object",
            "properties": {
                "extraTime": {
                    "type": "string"
                }
            }
        },
        "main.Match": {
            "description": "Modelo que contiene la información básica de un partido",
            "type": "object",
            "properties": {
                "awayTeam": {
                    "type": "string"
                },
                "extraTime": {
                    "type": "string"
                },
                "homeTeam": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "matchDate": {
                    "type": "string"
                }
            }
        }
    }
}